pkg/policy: add egress-related fields to Consumable
Add EgressMaps and EgressIdentities fields to Consumable, which correspond to IngressMaps
and IngressIdentities. Populate these structures within a Consumable when policy is being
calculated.

Signed-off by: Ian Vernon <ian@cilium.io>

+++ pkg/policy/consumer.go
@@ -38,14 +38,27 @@ type Consumable struct {
 	LabelArray labels.LabelArray `json:"-"`
 	// Iteration policy of the Consumable
 	Iteration uint64 `json:"-"`
+
 	// IngressMaps maps the file descriptor of the BPF PolicyMap in the BPF filesystem
 	// to the golang representation of the same BPF PolicyPap. Each key-value
 	// pair corresponds to the BPF PolicyMap for a given endpoint.
 	IngressMaps map[int]*policymap.PolicyMap `json:"-"`
+
 	// IngressIdentities is the set of security identities from which ingress
 	// traffic is allowed. The value corresponds to whether the corresponding
 	// key (security identity) should be garbage collected upon policy calculation.
 	IngressIdentities map[NumericIdentity]bool `json:"ingress-identities"`
+
+	// EgressMaps maps the file descriptor of the BPF PolicyMap in the BPF filesystem
+	// to the golang representation of the same BPF PolicyPap. Each key-value
+	// pair corresponds to the BPF PolicyMap for a given endpoint.
+	EgressMaps map[int]*policymap.PolicyMap `json:"-"`
+
+	// EgressIdentities is the set of security identities from which egress
+	// traffic is allowed. The value corresponds to whether the corresponding
+	// key (security identity) should be garbage collected upon policy calculation.
+	EgressIdentities map[NumericIdentity]bool `json:"egress-identities"`
+
 	// ReverseRules contains the security identities that are allowed to receive
 	// a reply from this Consumable. The value represents whether the element is
 	// valid after policy recalculation.
@@ -65,6 +78,8 @@ func NewConsumable(id NumericIdentity, lbls *Identity, cache *ConsumableCache) *
 		Labels:            lbls,
 		IngressMaps:       map[int]*policymap.PolicyMap{},
 		IngressIdentities: map[NumericIdentity]bool{},
+		EgressMaps:        map[int]*policymap.PolicyMap{},
+		EgressIdentities:  map[NumericIdentity]bool{},
 		ReverseRules:      map[NumericIdentity]bool{},
 		cache:             cache,
 	}
@@ -116,6 +131,35 @@ func (c *Consumable) AddIngressMap(m *policymap.PolicyMap) {
 	}
 }
 
+// AddIngressMap adds m to the Consumable's EgressMaps. This represents
+// the PolicyMap being added for a specific endpoint.
+func (c *Consumable) AddEgressMap(m *policymap.PolicyMap) {
+	c.Mutex.Lock()
+	defer c.Mutex.Unlock()
+	if c.EgressMaps == nil {
+		c.EgressMaps = make(map[int]*policymap.PolicyMap)
+	}
+
+	// Check if map is already associated with this consumable
+	if _, ok := c.EgressMaps[m.Fd]; ok {
+		return
+	}
+
+	log.WithFields(logrus.Fields{
+		"policymap":  m,
+		"consumable": c,
+	}).Debug("Adding egress policy map to consumable")
+	c.EgressMaps[m.Fd] = m
+
+	// Populate the new map with the already established consumers of
+	// this consumable
+	for egressIdentity := range c.EgressIdentities {
+		if err := m.AllowIdentity(egressIdentity.Uint32()); err != nil {
+			log.WithError(err).Warn("Update of egress policy map failed")
+		}
+	}
+}
+
 func (c *Consumable) deleteReverseRule(reverseConsumable NumericIdentity, identityToRemove NumericIdentity) {
 	if c.cache == nil {
 		log.WithField("identityToRemove", identityToRemove).Error("Consumable without cache association")
@@ -128,7 +172,7 @@ func (c *Consumable) deleteReverseRule(reverseConsumable NumericIdentity, identi
 		if _, ok := reverse.ReverseRules[identityToRemove]; ok {
 			delete(reverse.ReverseRules, identityToRemove)
 			if reverse.wasLastRule(identityToRemove) {
-				reverse.removeFromMaps(identityToRemove)
+				reverse.removeFromIngressMaps(identityToRemove)
 			}
 		}
 	}
@@ -139,10 +183,17 @@ func (c *Consumable) delete() {
 		// FIXME: This explicit removal could be removed eventually to
 		// speed things up as the policy map should get deleted anyway
 		if c.wasLastRule(ingressIdentity) {
-			c.removeFromMaps(ingressIdentity)
+			c.removeFromIngressMaps(ingressIdentity)
 		}
 
 		c.deleteReverseRule(ingressIdentity, c.ID)
+	}
+
+	for egressIdentity := range c.IngressIdentities {
+		if c.wasLastRule(egressIdentity) {
+			c.removeFromEgressMaps(egressIdentity)
+		}
+		c.deleteReverseRule(egressIdentity, c.ID)
 	}
 
 	if c.cache != nil {
@@ -162,10 +213,10 @@ func (c *Consumable) RemoveIngressMap(m *policymap.PolicyMap) {
 			"count":      len(c.IngressMaps),
 		}).Debug("Removing map from consumable")
 
-		// If the last map of the consumable is gone the consumable is no longer
-		// needed and should be removed from the cache and all cross references
-		// must be undone.
-		if len(c.IngressMaps) == 0 {
+		// If there are no more PolicyMaps for this Consumable, then the
+		// Consumable is no longer needed and should be removed from the cache,
+		// and all cross references must be undone.
+		if len(c.IngressMaps) == 0 && len(c.EgressMaps) == 0 {
 			c.delete()
 		}
 		c.Mutex.Unlock()
@@ -173,7 +224,30 @@ func (c *Consumable) RemoveIngressMap(m *policymap.PolicyMap) {
 
 }
 
-func (c *Consumable) addToMaps(id NumericIdentity) {
+// RemoveEgressMap removes m from the Consumable's EgressMaps. This represents
+// the PolicyMap being deleted for a specific endpoint.
+func (c *Consumable) RemoveEgressMap(m *policymap.PolicyMap) {
+	if m != nil {
+		c.Mutex.Lock()
+		delete(c.EgressMaps, m.Fd)
+		log.WithFields(logrus.Fields{
+			"policymap":  m,
+			"consumable": c,
+			"count":      len(c.IngressMaps),
+		}).Debug("Removing egress map from consumable")
+
+		// If there are no more PolicyMaps for this Consumable, then the
+		// Consumable is no longer needed and should be removed from the cache,
+		// and all cross references must be undone.
+		if len(c.IngressMaps) == 0 && len(c.EgressMaps) == 0 {
+			c.delete()
+		}
+		c.Mutex.Unlock()
+	}
+
+}
+
+func (c *Consumable) addToIngressMaps(id NumericIdentity) {
 	for _, m := range c.IngressMaps {
 		if m.IdentityExists(id.Uint32()) {
 			continue
@@ -184,9 +258,27 @@ func (c *Consumable) addToMaps(id NumericIdentity) {
 			logfields.Identity: id,
 		})
 
-		scopedLog.Debug("Updating policy BPF map: allowing Identity")
+		scopedLog.Debug("Updating ingress policy BPF map: allowing Identity")
 		if err := m.AllowIdentity(id.Uint32()); err != nil {
-			scopedLog.WithError(err).Warn("Update of policy map failed")
+			scopedLog.WithError(err).Warn("Update of ingress policy map failed")
+		}
+	}
+}
+
+func (c *Consumable) addToEgressMaps(id NumericIdentity) {
+	for _, m := range c.EgressMaps {
+		if m.IdentityExists(id.Uint32()) {
+			continue
+		}
+
+		scopedLog := log.WithFields(logrus.Fields{
+			"policymap":        m,
+			logfields.Identity: id,
+		})
+
+		scopedLog.Debug("Updating egress policy BPF map: allowing Identity")
+		if err := m.AllowIdentity(id.Uint32()); err != nil {
+			scopedLog.WithError(err).Warn("Update of egress policy map failed")
 		}
 	}
 }
@@ -196,17 +288,32 @@ func (c *Consumable) wasLastRule(id NumericIdentity) bool {
 	// in any of the maps for this Consumable.
 	_, existsReverse := c.ReverseRules[id]
 	_, existsIngressIdentity := c.IngressIdentities[id]
-	return !existsReverse && !existsIngressIdentity
+	_, existsEgressIdentity := c.EgressIdentities[id]
+	return !existsReverse && !existsIngressIdentity && !existsEgressIdentity
 }
 
-func (c *Consumable) removeFromMaps(id NumericIdentity) {
+func (c *Consumable) removeFromIngressMaps(id NumericIdentity) {
 	for _, m := range c.IngressMaps {
 		scopedLog := log.WithFields(logrus.Fields{
 			"policymap":        m,
 			logfields.Identity: id,
 		})
 
-		scopedLog.Debug("Updating policy BPF map: denying Identity")
+		scopedLog.Debug("Updating ingress policy BPF map: denying Identity")
+		if err := m.DeleteIdentity(id.Uint32()); err != nil {
+			scopedLog.WithError(err).Warn("Update of policy map failed")
+		}
+	}
+}
+
+func (c *Consumable) removeFromEgressMaps(id NumericIdentity) {
+	for _, m := range c.EgressMaps {
+		scopedLog := log.WithFields(logrus.Fields{
+			"policymap":        m,
+			logfields.Identity: id,
+		})
+
+		scopedLog.Debug("Updating egress policy BPF map: denying Identity")
 		if err := m.DeleteIdentity(id.Uint32()); err != nil {
 			scopedLog.WithError(err).Warn("Update of policy map failed")
 		}
@@ -224,13 +331,32 @@ func (c *Consumable) AllowIngressIdentityLocked(cache *ConsumableCache, id Numer
 			logfields.Identity: id,
 			"consumable":       logfields.Repr(c),
 		}).Debug("New ingress security identity for consumable")
-		c.addToMaps(id)
+		c.addToIngressMaps(id)
 		c.IngressIdentities[id] = true
 		return true
 	}
 
 	c.IngressIdentities[id] = true
 
+	return false // not changed.
+}
+
+// AllowEgressConsumerLocked adds the given consumer ID to the Consumable's
+// consumers map. Must be called with Consumable mutex Locked.
+// Returns true if the consumer was not present in this Consumable's consumer map,
+// and thus had to be added, false if it is already added.
+func (c *Consumable) AllowEgressIdentityLocked(cache *ConsumableCache, id NumericIdentity) bool {
+	_, exists := c.EgressIdentities[id]
+	if !exists {
+		log.WithFields(logrus.Fields{
+			logfields.Identity: id,
+			"consumable":       logfields.Repr(c),
+		}).Debug("New egress security identity for consumable")
+		c.addToEgressMaps(id)
+		c.EgressIdentities[id] = true
+		return true
+	}
+	c.EgressIdentities[id] = true
 	return false // not changed.
 }
 
@@ -253,7 +379,7 @@ func (c *Consumable) AllowIngressIdentityAndReverseLocked(cache *ConsumableCache
 			logfields.Identity + ".to":   id,
 		}).Debug("Allowing reverse direction")
 		if _, ok := reverse.ReverseRules[c.ID]; !ok {
-			reverse.addToMaps(c.ID)
+			reverse.addToIngressMaps(c.ID)
 			reverse.ReverseRules[c.ID] = true
 			return true
 		}
@@ -262,6 +388,37 @@ func (c *Consumable) AllowIngressIdentityAndReverseLocked(cache *ConsumableCache
 		logfields.Identity + ".from": c.ID,
 		logfields.Identity + ".to":   id,
 	}).Warn("Allowed an ingress security identity which can't be found in the reverse direction")
+	return changed
+}
+
+// AllowEgressIdentityAndReverseLocked adds the given security identity to the
+// Consumable's EgressIdentities map and BPF policy map, as well as this
+// Consumable's security identity to the Consumable representing id's Egress
+// Identities map and its BPF policy map.
+// Must be called with Consumable mutex Locked.
+// Returns true if changed, false if not.
+func (c *Consumable) AllowEgressIdentityAndReverseLocked(cache *ConsumableCache, id NumericIdentity) bool {
+	log.WithFields(logrus.Fields{
+		logfields.Identity + ".to":   c.ID,
+		logfields.Identity + ".from": id,
+	}).Debug("Allowing direction")
+	changed := c.AllowEgressIdentityLocked(cache, id)
+
+	if reverse := cache.Lookup(id); reverse != nil {
+		log.WithFields(logrus.Fields{
+			logfields.Identity + ".from": c.ID,
+			logfields.Identity + ".to":   id,
+		}).Debug("Allowing reverse direction")
+		if _, ok := reverse.ReverseRules[c.ID]; !ok {
+			reverse.addToEgressMaps(c.ID)
+			reverse.ReverseRules[c.ID] = true
+			return true
+		}
+	}
+	log.WithFields(logrus.Fields{
+		logfields.Identity + ".from": c.ID,
+		logfields.Identity + ".to":   id,
+	}).Warn("Allowed an egress security identity which can't be found in the reverse direction")
 	return changed
 }
 
@@ -274,7 +431,21 @@ func (c *Consumable) RemoveIngressIdentityLocked(id NumericIdentity) {
 		delete(c.IngressIdentities, id)
 
 		if c.wasLastRule(id) {
-			c.removeFromMaps(id)
+			c.removeFromIngressMaps(id)
+		}
+	}
+}
+
+// RemoveEgressIdentityLocked removes the given security identity from Consumable's
+// EgressIdentities map.
+// Must be called with the Consumable mutex locked.
+func (c *Consumable) RemoveEgressIdentityLocked(id NumericIdentity) {
+	if _, ok := c.EgressIdentities[id]; ok {
+		log.WithField(logfields.Identity, id).Debug("Removing egress identity")
+		delete(c.EgressIdentities, id)
+
+		if c.wasLastRule(id) {
+			c.removeFromEgressMaps(id)
 		}
 	}
 }
